backend/pyproject.toml
[tool.poetry]
name = "one-site-api"
version = "0.1.0"
description = "One Site API"
authors = ["Roba-V <roba@vrstu.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
pydantic-settings = "^2.4.0"
sqlmodel = "^0.0.21"
asyncpg = "^0.29.0"
pydantic = "^2.8.2"
psycopg2-binary = "^2.9.9"
greenlet = "^3.0.3"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.0"
isort = "^5.13.2"
mypy = "^1.11.1"
pre-commit = "^3.8.0"
pytest-asyncio = "^0.23.8"
aiosqlite = "^0.20.0"
httpx = "^0.27.0"
coverage = "^7.6.1"
docformatter = "^1.7.5"
parameterized = "^0.9.0"
types-passlib = "^1.7.7.20240819"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
exclude = ".venv"

[tool.docformatter]
in-place = true
pre-summary-newline = true
recursive = true
